made equipment overlay own class
lmk if there's any bugs
